pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "newsfeed-digest"
version = "0.1.0"
description = "Travel industry digest automation for growth/marketing teams"
requires-python = ">=3.10"
dependencies = [
  "feedparser==6.0.11",
  "requests==2.32.3",
  "python-dateutil==2.9.0.post0",
  "beautifulsoup4==4.12.3",
  "lxml==5.3.0",
  "jinja2==3.1.4",
  "python-dotenv==1.0.1",
  "openai==1.58.1",
  "tenacity==9.0.0",
  "pyyaml==6.0.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
newsfeed = "newsfeed.cli:main"

